CAT updates: Removed the admin IP stuff since the security groups are wide open now anyway. :o Fixed some typos in the security group configs.
rs_kubernetes.sh:
Fixed a problem where the kubectl commands were failing due to not
finding localhost:8080.
This was because things live on 6443 now. The kubeadm init command lists
a sequence of lines that need to be run after the init - this basically
creates a local kube config file which provides the right path, etc.